refactor(agents): extract agent liveness concern and client header builder
