+++ engines/collavre/app/models/collavre/user.rb
@@ -3,6 +3,7 @@ module Collavre
     self.table_name = "users"
 
     include HasInboxCreative
+    include AgentLiveness
 
     has_many :user_themes, class_name: "Collavre::UserTheme", dependent: :destroy
 
@@ -212,31 +213,6 @@ module Collavre
       return unless old_gateway_id
 
       agent_workspaces.where(agent_gateway_id: old_gateway_id).destroy_all
-    end
-
-    def claude_channel_agent?
-      llm_model == "claude-code"
-    end
-
-    def claude_channel_online?
-      claude_channel_agent? && AgentSubscription.live.where(agent_id: id).exists?
-    end
-
-    # A gateway-backed agent is reachable when its gateway's last readiness
-    # probe is recent, positive, and does not name this agent's own engine as
-    # logged out. Deliberately independent of chat presence: nobody has to have
-    # the creative open for the agent to be able to run.
-    def gateway_online?
-      return false unless cli_proxy_agent?
-
-      agent_gateway.health_serves_engine?(CliProxy::AdapterEngine.for_model(llm_model))
-    end
-
-    # Whether this agent can be dispatched to right now, for the two agent kinds
-    # that publish evidence either way. Agents on a hosted vendor API publish
-    # none, so they stay out of it rather than being asserted online.
-    def agent_online?
-      claude_channel_online? || gateway_online?
     end
 
     scope :ai_agents, -> { where.not(llm_vendor: [ nil, "" ]) }
